change constructor input types

+++ src/product/pusher.cpp
@@ -5,7 +5,7 @@
 #define OPEN_STATE 0 //put here for easier editing
 #define CLOSE_STATE 90
 
-pusher::pusher(int temp_photo_pin, int temp_servo_pin){
+pusher::pusher(uint8_t temp_photo_pin, uint8_t temp_servo_pin){
   photo_sensor = new sensor(temp_photo_pin);
   servo_sensor = new sensor(temp_servo_pin);
   the_servo.attach(servo_sensor.pin);
